+++ src/Renderer/RayTracer.cpp
@@ -331,7 +331,7 @@ namespace OfflineRenderer {
                 // glm::vec3 sampleCol = GetColor(ray, 0, brdfSamples, j);
 
                 glm::vec3 sampleCol = ray.direction;
-                color += sampleCol * numSamplesInv;
+                color += sampleCol * (float)numSamplesInv;
             }
             // if(y > 300)  std::cout << x << "  " << y <<  std::endl;
